interface/requirements.txt
anthropic==0.34.1
distro==1.9.0
Django==4.2.15
django-sslserver==0.22
djangorestframework==3.15.2
filelock==3.15.4
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.8
jiter==0.5.0
openai==1.42.0
packaging==24.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
SpeechRecognition==3.10.4
urllib3==2.2.2
whitenoise==6.7.0
